Improved dot system. Abstracts from firm Graph. Adds dependency to petgraph for custom graphs.
